+++ frontend/src/App.js
@@ -1,12 +1,12 @@
 
 import './App.css';
-import Header from './Header';
+import Home from './Components/Home';
 
 function App() {
 
   return (
     <div className="App">
-      <Header />
+      <Home />
     </div>
 
   );
