+++ frontend/src/Components/Home.js
@@ -1,0 +1,15 @@
+import React from 'react';
+import Header from './Header'
+import Pun from './Pun';
+
+
+function Home() {
+    return (
+        <div className="Home Page">
+            <Header />
+            <Pun />
+        </div>
+    );
+}
+
+export default Home;
